Set the expiration date for all ports which depend upon linux_base-fc4 to the EoL of the last 7.x release. The 7.x branch is the only supported FreeBSD version not yet at his EoL which needs a linux 2.4 based linuxulator environment (linux_base-f10 doesn't work there).
I didn't bump the portrevision, this is mostly done as a hint for ports
committers to tell them that at least the emulation@-owned ports need
to stay until then (even the forbidden ones). The dependent ports need to
go at the same time, but are free to be deleted before.

Discussed with:	bsam