Add the 0.16 branch of libtorrent-rasterbar
The most revolutionary aspect of the new branch is that it brings in
support for uTP.

This port also brings in a patch consisting of the changes from 0.16.0
release through local SVN revision 6850. These fixes include patches we
had in the 0.15 version of the port, as well as various other fixes that
were reported by me, and others post the 0.16.0 release.  These fixes
should be in the 0.16.1 release.