- New port math/octave-forge-java.
The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

This is java, and provides a Java interface with OO-like Java objects
manipulation.